Speed up tests by localizing resources
