+++ lib/constants/index.js
@@ -6,8 +6,7 @@ export const isServer = typeof window === 'undefined'
 /**
  * Logo
  */
-export const LOGO_URL =
-  'https://d2f1n6ed3ipuic.cloudfront.net/conveyal-128x128.png'
+export const LOGO_URL = '/img/conveyal-128x128.png'
 export const CB_HEX = '#2389c9'
 export const CB_DARK = '#185c86'
 export const CB_RGB = {r: 35, g: 137, b: 201}
